@@ -20,34 +20,41 @@
     "postbuild": "next-sitemap --config next-sitemap.config.js"
   },
   "dependencies": {
-    "axios": ">=0.21.1",
+    "axios": "^1.3.4",
+    "classnames": "^2.3.2",
+    "dayjs": "^1.11.7",
     "feed": "^4.2.2",
     "gitalk": "^1.7.2",
-    "next": "^12.1.0",
-    "notion-client": "^4.10.0",
-    "notion-utils": "^4.10.0",
+    "mermaid": "^10.0.2",
+    "next": "^13.2.4",
+    "notion-client": "^6.16.0",
+    "notion-utils": "^6.16.0",
     "preact": "^10.5.15",
-    "react": "^17.0.2",
-    "react-cusdis": "^2.0.1",
-    "react-dom": "^17.0.2",
-    "react-notion-x": "^4.11.0",
-    "use-ackee": "^3.0.0"
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-cusdis": "^2.1.3",
+    "react-dom": "^18.2.0",
+    "react-notion-x": "^6.16.0",
+    "react-tweet-embed": "^2.0.0",
+    "react-use": "^17.4.0",
+    "use-ackee": "^3.0.1"
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.0",
-    "eslint": "<8.0.0",
-    "eslint-config-next": "^12.0.3",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.1",
-    "eslint-plugin-react": "^7.26.1",
-    "next-sitemap": "^1.6.203",
-    "postcss": "^8.3.11",
-    "tailwindcss": "^2.2.19"
+    "autoprefixer": "^10.4.14",
+    "eslint": "^8.36.0",
+    "eslint-config-next": "^13.2.4",
+    "next-sitemap": "^4.0.5",
+    "postcss": "^8.4.21",
+    "tailwindcss": "^3.3.1"
   },
   "bugs": {
     "url": "https://github.com/craigary/nobelium/issues",
     "email": "i@craigary.net"
+  },
+  "pnpm": {
+    "patchedDependencies": {
+      "react-notion-x@6.16.0": "patches/react-notion-x@6.16.0.patch",
+      "notion-utils@6.16.0": "patches/notion-utils@6.16.0.patch"
+    }
   }
 }
