@@ -1,8 +1,9 @@
-import { fetchCusdisLang } from '@/lib/cusdisLang'
-import BLOG from '@/blog.config'
-import dynamic from 'next/dynamic'
-import { useRouter } from 'next/router'
 import 'gitalk/dist/gitalk.css'
+import { useRouter } from 'next/router'
+import dynamic from 'next/dynamic'
+import cn from 'classnames'
+import { fetchCusdisLang } from '@/lib/cusdisLang'
+import { useConfig } from '@/lib/config'
 
 const GitalkComponent = dynamic(
   () => {
@@ -25,8 +26,17 @@ const CusdisComponent = dynamic(
 
 const Comments = ({ frontMatter }) => {
   const router = useRouter()
+  const BLOG = useConfig()
+
+  const fullWidth = frontMatter.fullWidth ?? false
+
   return (
-    <div>
+    <div
+      className={cn(
+        'px-4 font-medium text-gray-500 dark:text-gray-400 my-5',
+        fullWidth ? 'md:px-24' : 'mx-auto max-w-2xl',
+      )}
+    >
       {BLOG.comment && BLOG.comment.provider === 'gitalk' && (
         <GitalkComponent
           options={{
@@ -46,7 +56,7 @@ const Comments = ({ frontMatter }) => {
       )}
       {BLOG.comment && BLOG.comment.provider === 'cusdis' && (
         <CusdisComponent
-        lang={fetchCusdisLang()}
+          lang={fetchCusdisLang(BLOG.lang)}
           attrs={{
             host: BLOG.comment.cusdisConfig.host,
             appId: BLOG.comment.cusdisConfig.appId,
