@@ -1,11 +1,14 @@
 import Header from '@/components/Header'
 import Footer from '@/components/Footer'
-import BLOG from '@/blog.config'
+import { useConfig } from '@/lib/config'
 import Head from 'next/head'
 import PropTypes from 'prop-types'
+import cn from 'classnames'
 // import BlogPost from './BlogPost'
 
 const Container = ({ children, layout, fullWidth, ...customMeta }) => {
+  const BLOG = useConfig()
+
   const url = BLOG.path.length ? `${BLOG.link}/${BLOG.path}` : BLOG.link
   const meta = {
     title: BLOG.title,
@@ -56,26 +59,24 @@ const Container = ({ children, layout, fullWidth, ...customMeta }) => {
           <>
             <meta
               property="article:published_time"
-              content={meta.date || meta.createdTime}
+              content={meta.date}
             />
             <meta property="article:author" content={BLOG.author} />
           </>
         )}
       </Head>
       <div
-        className={`wrapper ${
-          BLOG.font === 'serif' ? 'font-serif' : 'font-sans'
-        }`}
+        className={`wrapper ${BLOG.font === 'serif' ? 'font-serif' : 'font-sans'
+          }`}
       >
         <Header
           navBarTitle={layout === 'blog' ? meta.title : null}
           fullWidth={fullWidth}
         />
-        <main
-          className={`m-auto flex-grow w-full transition-all ${
-            !fullWidth ? 'max-w-2xl px-4' : 'px-4 md:px-24'
-          }`}
-        >
+        <main className={cn(
+          'flex-grow transition-all',
+          layout !== 'blog' && ['self-center px-4', fullWidth ? 'md:px-24' : 'w-full max-w-2xl']
+        )}>
           {children}
         </main>
         <Footer fullWidth={fullWidth} />
