@@ -1,6 +1,8 @@
-import BLOG from '@/blog.config'
+import { useConfig } from '@/lib/config'
 import Vercel from '@/components/Vercel'
 const Footer = ({ fullWidth }) => {
+  const BLOG = useConfig()
+
   const d = new Date()
   const y = d.getFullYear()
   const from = +BLOG.since
