@@ -1,48 +1,35 @@
 import Script from 'next/script'
-import BLOG from '@/blog.config'
+import { useConfig } from '@/lib/config'
 
-const Scripts = () => (
-  <>
-    {BLOG.analytics && BLOG.analytics.provider === 'ackee' && (
-      <Script
-        src={BLOG.analytics.ackeeConfig.tracker}
-        data-ackee-server={BLOG.analytics.ackeeConfig.dataAckeeServer}
-        data-ackee-domain-id={BLOG.analytics.ackeeConfig.domainId}
-      />
-    )}
-    {/* {BLOG.autoCollapsedNavBar === true && (
-      <Script strategy="lazyOnload">
-        {`var windowTop=0;
-          function scrollTrigger(){
-              let scrollS = window.scrollY;
-              let nav = document.querySelector('.sticky-nav');
-              if(scrollS >= windowTop){
-                  nav.style.opacity = 0;
-                  windowTop = scrollS;
-              }else{
-                  nav.style.opacity = 1;
-                  windowTop = scrollS;
-              }
-          };
-          window.addEventListener('scroll',scrollTrigger);`}
-      </Script>
-    )} */}
-    {BLOG.analytics && BLOG.analytics.provider === 'ga' && (
-      <>
+const Scripts = () => {
+  const BLOG = useConfig()
+
+  return (
+    <>
+      {BLOG.analytics && BLOG.analytics.provider === 'ackee' && (
         <Script
-          src={`https://www.googletagmanager.com/gtag/js?id=${BLOG.analytics.gaConfig.measurementId}`}
+          src={BLOG.analytics.ackeeConfig.tracker}
+          data-ackee-server={BLOG.analytics.ackeeConfig.dataAckeeServer}
+          data-ackee-domain-id={BLOG.analytics.ackeeConfig.domainId}
         />
-        <Script strategy="lazyOnload" id="ga">
-          {`window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${BLOG.analytics.gaConfig.measurementId}', {
-              page_path: window.location.pathname,
-            });`}
-        </Script>
-      </>
-    )}
-  </>
-)
+      )}
+      {BLOG.analytics && BLOG.analytics.provider === 'ga' && (
+        <>
+          <Script
+            src={`https://www.googletagmanager.com/gtag/js?id=${BLOG.analytics.gaConfig.measurementId}`}
+          />
+          <Script strategy="lazyOnload" id="ga">
+            {`window.dataLayer = window.dataLayer || [];
+              function gtag(){dataLayer.push(arguments);}
+              gtag('js', new Date());
+              gtag('config', '${BLOG.analytics.gaConfig.measurementId}', {
+                page_path: window.location.pathname,
+              });`}
+          </Script>
+        </>
+      )}
+    </>
+  )
+}
 
 export default Scripts
