@@ -1,6 +1,6 @@
-const BLOG = require('../blog.config')
+const { config: BLOG } = require('../lib/server/config')
 
-module.exports = function () {
+function cjk() {
   switch (BLOG.lang.toLowerCase()) {
     case 'zh-cn':
     case 'zh-sg':
@@ -19,3 +19,5 @@ module.exports = function () {
       return null
   }
 }
+
+module.exports = cjk
