@@ -1,5 +1,3 @@
-import BLOG from '@/blog.config'
-
 const cusdisI18n = [
   'ca',
   'en',
@@ -15,10 +13,8 @@ const cusdisI18n = [
   'zh-tw'
 ]
 
-const loweredLang = BLOG.lang.toLowerCase()
-
-export const fetchCusdisLang = () => {
-  if (BLOG.comment.provider !== 'cusdis') return null
+export const fetchCusdisLang = lang => {
+  const loweredLang = lang.toLowerCase()
   if (loweredLang.startsWith('zh')) {
     return (
       cusdisI18n.find(i => loweredLang === i.toLocaleLowerCase()) ?? 'zh-cn'
@@ -26,7 +22,7 @@ export const fetchCusdisLang = () => {
   } else {
     return (
       cusdisI18n.find(i =>
-        BLOG.lang.toLowerCase().startsWith(i.toLowerCase())
+        loweredLang.startsWith(i.toLowerCase())
       ) ?? 'en'
     )
   }
