@@ -1,9 +1,6 @@
-import BLOG from '@/blog.config'
-export const GA_TRACKING_ID = BLOG.analytics.gaConfig.measurementId
-
 // https://developers.google.com/analytics/devguides/collection/gtagjs/pages
-export const pageview = url => {
-  window.gtag('config', GA_TRACKING_ID, {
+export const pageview = (gTag, url) => {
+  window.gtag('config', gTag, {
     page_path: url
   })
 }
