@@ -1,12 +1,12 @@
-import { useContext, createContext } from 'react'
-import { fetchLocaleLang } from '@/lib/lang'
+import { createContext, useContext } from 'react'
 
-const locale = fetchLocaleLang()
-const LocaleContext = createContext()
+const LocaleContext = createContext(undefined)
 
-export function LocaleProvider ({ children }) {
+export function LocaleProvider ({ value, children }) {
   return (
-    <LocaleContext.Provider value={locale}>{children}</LocaleContext.Provider>
+    <LocaleContext.Provider value={value}>
+      {children}
+    </LocaleContext.Provider>
   )
 }
 
