@@ -1,6 +1,8 @@
-import BLOG from '@/blog.config'
-import { NotionAPI } from 'notion-client'
+import { config as BLOG } from '@/lib/server/config'
+
 import { idToUuid } from 'notion-utils'
+import dayjs from 'dayjs'
+import api from '@/lib/server/notion-api'
 import getAllPageIds from './getAllPageIds'
 import getPageProperties from './getPageProperties'
 import filterPublishedPosts from './filterPublishedPosts'
@@ -9,12 +11,10 @@ import filterPublishedPosts from './filterPublishedPosts'
  * @param {{ includePages: boolean }} - false: posts only / true: include pages
  */
 export async function getAllPosts ({ includePages = false }) {
-  let id = BLOG.notionPageId
-  const authToken = BLOG.notionAccessToken || null
-  const api = new NotionAPI({ authToken })
+  const id = idToUuid(process.env.NOTION_PAGE_ID)
+
   const response = await api.getPage(id)
 
-  id = idToUuid(id)
   const collection = Object.values(response.collection)[0]?.value
   const collectionQuery = response.collection_query
   const block = response.block
@@ -37,11 +37,14 @@ export async function getAllPosts ({ includePages = false }) {
       const id = pageIds[i]
       const properties = (await getPageProperties(id, block, schema)) || null
 
-      // Add fullwidth, createdtime to properties
-      properties.createdTime = new Date(
-        block[id].value?.created_time
-      ).toString()
+      // Add fullwidth to properties
       properties.fullWidth = block[id].value?.format?.page_full_width ?? false
+      // Convert date (with timezone) to unix milliseconds timestamp
+      properties.date = (
+        properties.date?.start_date
+          ? dayjs.tz(properties.date?.start_date)
+          : dayjs(block[id].value?.created_time)
+      ).valueOf()
 
       data.push(properties)
     }
@@ -51,11 +54,7 @@ export async function getAllPosts ({ includePages = false }) {
 
     // Sort by date
     if (BLOG.sortByDate) {
-      posts.sort((a, b) => {
-        const dateA = new Date(a?.date?.start_date || a.createdTime)
-        const dateB = new Date(b?.date?.start_date || b.createdTime)
-        return dateB - dateA
-      })
+      posts.sort((a, b) => b.date - a.date)
     }
     return posts
   }
