@@ -1,8 +1,7 @@
 import { getTextContent, getDateValue } from 'notion-utils'
-import { NotionAPI } from 'notion-client'
+import api from '@/lib/server/notion-api'
 
-async function getPageProperties (id, block, schema, authToken) {
-  const api = new NotionAPI({ authToken })
+async function getPageProperties (id, block, schema) {
   const rawProperties = Object.entries(block?.[id]?.value?.properties || [])
   const excludeProperties = ['date', 'select', 'multi_select', 'person']
   const properties = {}
