@@ -1,9 +1,6 @@
-import BLOG from '@/blog.config'
-import { NotionAPI } from 'notion-client'
+import api from '@/lib/server/notion-api'
 
 export async function getPostBlocks (id) {
-  const authToken = BLOG.notionAccessToken || null
-  const api = new NotionAPI({ authToken })
   const pageBlock = await api.getPage(id)
   return pageBlock
 }
