@@ -1,50 +1,44 @@
 import { Feed } from 'feed'
-import BLOG from '@/blog.config'
 import ReactDOMServer from 'react-dom/server'
+import { clientConfig, config } from '@/lib/server/config'
 import { getPostBlocks } from '@/lib/notion'
-import { NotionRenderer, Equation, Code, Collection, CollectionRow } from 'react-notion-x'
-
-const mapPageUrl = id => 'https://www.notion.so/' + id.replace(/-/g, '')
+import { ConfigProvider } from '@/lib/config'
+import NotionRenderer from '@/components/NotionRenderer'
 
 const createFeedContent = async post => {
-  const content = ReactDOMServer.renderToString(<NotionRenderer
-    recordMap={await getPostBlocks(post.id)}
-    components={{
-      equation: Equation,
-      code: Code,
-      collection: Collection,
-      collectionRow: CollectionRow
-    }}
-    mapPageUrl={mapPageUrl}
-  />)
+  const content = ReactDOMServer.renderToString(
+    <ConfigProvider value={clientConfig}>
+      <NotionRenderer recordMap={await getPostBlocks(post.id)} />
+    </ConfigProvider>
+  )
   const regexExp = /<div class="notion-collection-row"><div class="notion-collection-row-body"><div class="notion-collection-row-property"><div class="notion-collection-column-title"><svg.*?class="notion-collection-column-title-icon">.*?<\/svg><div class="notion-collection-column-title-body">.*?<\/div><\/div><div class="notion-collection-row-value">.*?<\/div><\/div><\/div><\/div>/g
   return content.replace(regexExp, '')
 }
 
-export async function generateRss (posts) {
+export async function generateRss(posts) {
   const year = new Date().getFullYear()
   const feed = new Feed({
-    title: BLOG.title,
-    description: BLOG.description,
-    id: `${BLOG.link}/${BLOG.path}`,
-    link: `${BLOG.link}/${BLOG.path}`,
-    language: BLOG.lang,
-    favicon: `${BLOG.link}/favicon.svg`,
-    copyright: `All rights reserved ${year}, ${BLOG.author}`,
+    title: config.title,
+    description: config.description,
+    id: `${config.link}/${config.path}`,
+    link: `${config.link}/${config.path}`,
+    language: config.lang,
+    favicon: `${config.link}/favicon.svg`,
+    copyright: `All rights reserved ${year}, ${config.author}`,
     author: {
-      name: BLOG.author,
-      email: BLOG.email,
-      link: BLOG.link
+      name: config.author,
+      email: config.email,
+      link: config.link
     }
   })
   for (const post of posts) {
     feed.addItem({
       title: post.title,
-      id: `${BLOG.link}/${post.slug}`,
-      link: `${BLOG.link}/${post.slug}`,
+      id: `${config.link}/${post.slug}`,
+      link: `${config.link}/${post.slug}`,
       description: post.summary,
       content: await createFeedContent(post),
-      date: new Date(post?.date?.start_date || post.createdTime)
+      date: new Date(post.date)
     })
   }
   return feed.atom1()
