@@ -1,9 +1,10 @@
-const BLOG = require('./blog.config')
+const { config }  = require('./lib/server/config')
 
 module.exports = {
-  siteUrl: BLOG.link,
+  siteUrl: config.link,
   generateRobotsTxt: true,
-  sitemapSize: 7000
+  sitemapSize: 7000,
+  generateIndexSitemap: false
   // ...other options
   // https://github.com/iamvishnusankar/next-sitemap#configuration-options
 }
