@@ -1,17 +1,11 @@
 module.exports = {
-  webpack5: true,
   images: {
     domains: ['gravatar.com']
   },
   eslint: {
-    dirs: [
-      'components',
-      'layouts',
-      'lib',
-      'pages'
-    ]
+    // dirs: ['components', 'layouts', 'lib', 'pages']
   },
-  async headers () {
+  async headers() {
     return [
       {
         source: '/:path*{/}?',
@@ -24,15 +18,16 @@ module.exports = {
       }
     ]
   },
-  webpack: (config, { dev, isServer }) => {
-    // Replace React with Preact only in client production build
-    if (!dev && !isServer) {
-      Object.assign(config.resolve.alias, {
-        react: 'preact/compat',
-        'react-dom/test-utils': 'preact/test-utils',
-        'react-dom': 'preact/compat'
-      })
-    }
-    return config
-  }
+  transpilePackages: ['dayjs']
+  // webpack: (config, { dev, isServer }) => {
+  //   // Replace React with Preact only in client production build
+  //   if (!dev && !isServer) {
+  //     Object.assign(config.resolve.alias, {
+  //       react: 'preact/compat',
+  //       'react-dom/test-utils': 'preact/test-utils',
+  //       'react-dom': 'preact/compat'
+  //     })
+  //   }
+  //   return config
+  // }
 }
