@@ -1,24 +1,79 @@
-import Layout from '@/layouts/layout'
-import { getAllPosts, getPostBlocks } from '@/lib/notion'
-import BLOG from '@/blog.config'
-import { createHash } from 'crypto'
+import { clientConfig } from '@/lib/server/config'
 
-const BlogPost = ({ post, blockMap, emailHash }) => {
-  if (!post) return null
+import { useRouter } from 'next/router'
+import cn from 'classnames'
+import { getAllPosts, getPostBlocks } from '@/lib/notion'
+import { useLocale } from '@/lib/locale'
+import { useConfig } from '@/lib/config'
+import { createHash } from 'crypto'
+import Container from '@/components/Container'
+import Post from '@/components/Post'
+import Comments from '@/components/Comments'
+
+export default function BlogPost ({ post, blockMap, emailHash }) {
+  const router = useRouter()
+  const BLOG = useConfig()
+  const locale = useLocale()
+
+  // TODO: It would be better to render something
+  if (router.isFallback) return null
+
+  const fullWidth = post.fullWidth ?? false
+
   return (
-    <Layout
-      blockMap={blockMap}
-      frontMatter={post}
-      emailHash={emailHash}
-      fullWidth={post.fullWidth}
-    />
+    <Container
+      layout="blog"
+      title={post.title}
+      description={post.summary}
+      slug={post.slug}
+      // date={new Date(post.publishedAt).toISOString()}
+      type="article"
+      fullWidth={fullWidth}
+    >
+      <Post
+        post={post}
+        blockMap={blockMap}
+        emailHash={emailHash}
+        fullWidth={fullWidth}
+      />
+
+      {/* Back and Top */}
+      <div
+        className={cn(
+          'px-4 flex justify-between font-medium text-gray-500 dark:text-gray-400 my-5',
+          fullWidth ? 'md:px-24' : 'mx-auto max-w-2xl'
+        )}
+      >
+        <a>
+          <button
+            onClick={() => router.push(BLOG.path || '/')}
+            className="mt-2 cursor-pointer hover:text-black dark:hover:text-gray-100"
+          >
+            ← {locale.POST.BACK}
+          </button>
+        </a>
+        <a>
+          <button
+            onClick={() => window.scrollTo({
+              top: 0,
+              behavior: 'smooth'
+            })}
+            className="mt-2 cursor-pointer hover:text-black dark:hover:text-gray-100"
+          >
+            ↑ {locale.POST.TOP}
+          </button>
+        </a>
+      </div>
+
+      <Comments frontMatter={post} />
+    </Container>
   )
 }
 
 export async function getStaticPaths () {
   const posts = await getAllPosts({ includePages: true })
   return {
-    paths: posts.map(row => `${BLOG.path}/${row.slug}`),
+    paths: posts.map(row => `${clientConfig.path}/${row.slug}`),
     fallback: true
   }
 }
@@ -26,9 +81,12 @@ export async function getStaticPaths () {
 export async function getStaticProps ({ params: { slug } }) {
   const posts = await getAllPosts({ includePages: true })
   const post = posts.find(t => t.slug === slug)
+
+  if (!post) return { notFound: true }
+
   const blockMap = await getPostBlocks(post.id)
   const emailHash = createHash('md5')
-    .update(BLOG.email)
+    .update(clientConfig.email)
     .digest('hex')
     .trim()
     .toLowerCase()
@@ -38,5 +96,3 @@ export async function getStaticProps ({ params: { slug } }) {
     revalidate: 1
   }
 }
-
-export default BlogPost
