@@ -1,14 +1,16 @@
+import { clientConfig } from '@/lib/server/config'
+
 import Container from '@/components/Container'
 import BlogPost from '@/components/BlogPost'
 import Pagination from '@/components/Pagination'
 import { getAllPosts } from '@/lib/notion'
-import BLOG from '@/blog.config'
+import { useConfig } from '@/lib/config'
 
 export async function getStaticProps () {
   const posts = await getAllPosts({ includePages: false })
-  const postsToShow = posts.slice(0, BLOG.postsPerPage)
+  const postsToShow = posts.slice(0, clientConfig.postsPerPage)
   const totalPosts = posts.length
-  const showNext = totalPosts > BLOG.postsPerPage
+  const showNext = totalPosts > clientConfig.postsPerPage
   return {
     props: {
       page: 1, // current page is 1
@@ -19,9 +21,11 @@ export async function getStaticProps () {
   }
 }
 
-const blog = ({ postsToShow, page, showNext }) => {
+export default function Blog ({ postsToShow, page, showNext }) {
+  const { title, description } = useConfig()
+
   return (
-    <Container title={BLOG.title} description={BLOG.description}>
+    <Container title={title} description={description}>
       {postsToShow.map(post => (
         <BlogPost key={post.id} post={post} />
       ))}
@@ -29,5 +33,3 @@ const blog = ({ postsToShow, page, showNext }) => {
     </Container>
   )
 }
-
-export default blog
